Подготовка БД для избранных треков и интерактора

+++ app/src/main/java/com/example/playlistmaker/data/db/FavoriteTrackEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "favorite_tracks")
 data class FavoriteTrackEntity(
     @PrimaryKey
-    val trackId: Long,
+    val trackId: Int,
 
     val artworkUrl100: String,
     val trackName: String,
