+++ app/src/main/java/com/example/playlistmaker/data/dto/TrackDto.kt
@@ -1,7 +1,7 @@
 package com.example.playlistmaker.data.dto
 
 data class TrackDto(
-    val trackId: Long,
+    val trackId: Int,
     val trackName: String,
     val artistName: String,
     val trackTimeMillis: Long,
